requirements.txt
python-telegram-bot>=22.0
python-dotenv==1.0.0
Pillow>=10.0.0
